Extract builders from tests/all into their own file
tests/all.rs is big enough. Also add documentation.
